requirements.txt
django
djangorestframework
django-cors-headers
pillow
psycopg2-binary
django-soft-delete
gunicorn
drf-yasg
django-filter
django-admin-autocomplete-filter

